Linux HAL updates
1. The type correction was merged as part of
   https://egit.irs.uni-stuttgart.de/eive/fsfw/pulls/7 in the
   EIVE project. Quotation of PR

definition of getSpiParameters is `void getSpiParameters(spi::SpiModes& spiMode, uint32_t& spiSpeed, UncommonParameters* parameters = nullptr) const;`.

Here, size_t spiSpeed is passed, which implicitely gets converted to a temporary, which can not be bound to uint32_t& and, at least in gcc 9.3.0, leads to a compiler error.

2. Allow flushing the UART buffers

+++ hal/src/fsfw_hal/linux/spi/SpiComIF.cpp
@@ -82,7 +82,7 @@ ReturnValue_t SpiComIF::initializeInterface(CookieIF *cookie) {
         gpioComIF->pullHigh(gpioId);
     }
 
-    size_t spiSpeed = 0;
+    uint32_t spiSpeed = 0;
     spi::SpiModes spiMode = spi::SpiModes::MODE_0;
 
     SpiCookie::UncommonParameters params;
